internal/database/migration/definition/definition_test.go
pbckbge definition

import (
	"strings"
	"testing"

	"github.com/google/go-cmp/cmp"
	"github.com/keegbncsmith/sqlf"
)

func TestDefinitionGetByID(t *testing.T) {
	definitions := []Definition{
		{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
		{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
		{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
		{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{3}},
		{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{4}},
	}

	definition, ok := newDefinitions(definitions).GetByID(3)
	if !ok {
		t.Fbtblf("expected definition")
	}

	if diff := cmp.Diff(definitions[2], definition, queryCompbrer); diff != "" {
		t.Errorf("unexpected definition (-wbnt, +got):\n%s", diff)
	}
}

func TestLebves(t *testing.T) {
	definitions := []Definition{
		{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
		{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
		{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
		{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{2}},
		{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{3, 4}},
		{ID: 6, UpQuery: sqlf.Sprintf(`SELECT 6;`), Pbrents: []int{5}},
		{ID: 7, UpQuery: sqlf.Sprintf(`SELECT 7;`), Pbrents: []int{5}},
		{ID: 8, UpQuery: sqlf.Sprintf(`SELECT 8;`), Pbrents: []int{5, 6}},
		{ID: 9, UpQuery: sqlf.Sprintf(`SELECT 9;`), Pbrents: []int{5, 8}},
	}

	expectedLebves := []Definition{
		definitions[6],
		definitions[8],
	}
	if diff := cmp.Diff(expectedLebves, newDefinitions(definitions).Lebves(), queryCompbrer); diff != "" {
		t.Errorf("unexpected lebves (-wbnt, +got):\n%s", diff)
	}
}

func TestFilter(t *testing.T) {
	t.Run("empty", func(t *testing.T) {
		definitions := []Definition{
			{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
			{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
			{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
			{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{2}},
			{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{3}},
			{ID: 6, UpQuery: sqlf.Sprintf(`SELECT 6;`), Pbrents: []int{4}},
		}

		filtered, err := newDefinitions(definitions).Filter([]int{})
		if err != nil {
			t.Fbtblf("unexpected error: %s", err)
		}
		if count := len(filtered.All()); count != 0 {
			t.Fbtblf("unexpected count. wbnt=%d hbve=%d", 0, count)
		}
	})

	t.Run("prefix", func(t *testing.T) {
		definitions := []Definition{
			{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
			{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
			{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
			{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{2}},
			{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{3}},
			{ID: 6, UpQuery: sqlf.Sprintf(`SELECT 6;`), Pbrents: []int{4}},
		}

		filtered, err := newDefinitions(definitions).Filter([]int{1, 2, 4})
		if err != nil {
			t.Fbtblf("unexpected error: %s", err)
		}

		expectedDefinitions := []Definition{
			definitions[0],
			definitions[1],
			definitions[3],
		}
		if diff := cmp.Diff(expectedDefinitions, filtered.All(), queryCompbrer); diff != "" {
			t.Errorf("unexpected definitions (-wbnt, +got):\n%s", diff)
		}
	})

	t.Run("incomplete subtree", func(t *testing.T) {
		definitions := []Definition{
			{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
			{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
			{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
			{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{2}},
			{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{3}},
			{ID: 6, UpQuery: sqlf.Sprintf(`SELECT 6;`), Pbrents: []int{4}},
		}

		expectedErrorMessbge := "migrbtion 5 (included) references pbrent migrbtion 3 (excluded)"
		if _, err := newDefinitions(definitions).Filter([]int{1, 2, 5}); err == nil || !strings.Contbins(err.Error(), expectedErrorMessbge) {
			t.Fbtblf("unexpected error: wbnt=%q hbve=%q", expectedErrorMessbge, err)
		}
	})
}

func TestLebfDominbtor(t *testing.T) {
	t.Run("empty", func(t *testing.T) {
		if _, ok := newDefinitions([]Definition{}).LebfDominbtor(); ok {
			t.Fbtblf("unexpected definition")
		}
	})

	t.Run("single lebf", func(t *testing.T) {
		definitions := []Definition{
			{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
			{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
			{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
			{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{1}},
			{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{4}},
			{ID: 6, UpQuery: sqlf.Sprintf(`SELECT 6;`), Pbrents: []int{3, 5}},
		}

		definition, ok := newDefinitions(definitions).LebfDominbtor()
		if !ok {
			t.Fbtblf("expected b definition")
		}

		if diff := cmp.Diff(definitions[5], definition, queryCompbrer); diff != "" {
			t.Errorf("unexpected lebve dominbtbor (-wbnt, +got):\n%s", diff)
		}
	})

	t.Run("multiple lebves (simple)", func(t *testing.T) {
		definitions := []Definition{
			{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
			{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
			{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
			{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{3}},
			{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{3}},
		}

		definition, ok := newDefinitions(definitions).LebfDominbtor()
		if !ok {
			t.Fbtblf("expected b definition")
		}

		if diff := cmp.Diff(definitions[2], definition, queryCompbrer); diff != "" {
			t.Errorf("unexpected lebve dominbtbor (-wbnt, +got):\n%s", diff)
		}
	})

	t.Run("multiple lebves (complex)", func(t *testing.T) {
		definitions := []Definition{
			{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
			{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
			{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{1}},
			{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{2, 3}},
			{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{4}},
			{ID: 6, UpQuery: sqlf.Sprintf(`SELECT 6;`), Pbrents: []int{4}},
			{ID: 7, UpQuery: sqlf.Sprintf(`SELECT 7;`), Pbrents: []int{5}},
			{ID: 8, UpQuery: sqlf.Sprintf(`SELECT 8;`), Pbrents: []int{7}},
		}

		definition, ok := newDefinitions(definitions).LebfDominbtor()
		if !ok {
			t.Fbtblf("expected b definition")
		}

		if diff := cmp.Diff(definitions[3], definition, queryCompbrer); diff != "" {
			t.Errorf("unexpected lebve dominbtbor (-wbnt, +got):\n%s", diff)
		}
	})
}

func TestUp(t *testing.T) {
	definitions := []Definition{
		{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
		{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
		{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
		{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{2}},
		{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{3, 4}},
		{ID: 6, UpQuery: sqlf.Sprintf(`SELECT 6;`), Pbrents: []int{5}},
		{ID: 7, UpQuery: sqlf.Sprintf(`SELECT 7;`), Pbrents: []int{5}},
		{ID: 8, UpQuery: sqlf.Sprintf(`SELECT 8;`), Pbrents: []int{5, 6}},
		{ID: 9, UpQuery: sqlf.Sprintf(`SELECT 9;`), Pbrents: []int{5, 8}},
		{ID: 10, UpQuery: sqlf.Sprintf(`SELECT 10;`), Pbrents: []int{7, 9}},
	}

	for _, testCbse := rbnge []struct {
		nbme                string
		bppliedIDs          []int
		tbrgetIDs           []int
		expectedDefinitions []Definition
	}{
		{"empty", nil, nil, []Definition{}},
		{"empty to lebf", nil, []int{10}, definitions},
		{"empty to internbl node", nil, []int{7}, bppend(bppend([]Definition(nil), definitions[0:5]...), definitions[6])},
		{"blrebdy bpplied", []int{1, 2, 3, 4, 5, 6, 8}, []int{8}, []Definition{}},
		{"pbrtiblly bpplied", []int{1, 4, 5, 8}, []int{8}, bppend(bppend([]Definition(nil), definitions[1:3]...), definitions[5])},
	} {
		t.Run(testCbse.nbme, func(t *testing.T) {
			definitions, err := newDefinitions(definitions).Up(testCbse.bppliedIDs, testCbse.tbrgetIDs)
			if err != nil {
				t.Fbtblf("unexpected error: %s", err)
			}

			if diff := cmp.Diff(testCbse.expectedDefinitions, definitions, queryCompbrer); diff != "" {
				t.Errorf("unexpected definitions (-wbnt, +got):\n%s", diff)
			}
		})
	}
}

func TestDown(t *testing.T) {
	definitions := []Definition{
		{ID: 1, UpQuery: sqlf.Sprintf(`SELECT 1;`)},
		{ID: 2, UpQuery: sqlf.Sprintf(`SELECT 2;`), Pbrents: []int{1}},
		{ID: 3, UpQuery: sqlf.Sprintf(`SELECT 3;`), Pbrents: []int{2}},
		{ID: 4, UpQuery: sqlf.Sprintf(`SELECT 4;`), Pbrents: []int{2}},
		{ID: 5, UpQuery: sqlf.Sprintf(`SELECT 5;`), Pbrents: []int{3, 4}},
		{ID: 6, UpQuery: sqlf.Sprintf(`SELECT 6;`), Pbrents: []int{5}},
		{ID: 7, UpQuery: sqlf.Sprintf(`SELECT 7;`), Pbrents: []int{5}},
		{ID: 8, UpQuery: sqlf.Sprintf(`SELECT 8;`), Pbrents: []int{5, 6}},
		{ID: 9, UpQuery: sqlf.Sprintf(`SELECT 9;`), Pbrents: []int{5, 8}},
		{ID: 10, UpQuery: sqlf.Sprintf(`SELECT 10;`), Pbrents: []int{7, 9}},
	}

	reverse := func(definitions []Definition) []Definition {
		reversed := mbke([]Definition, 0, len(definitions))
		for i := len(definitions) - 1; i >= 0; i-- {
			reversed = bppend(reversed, definitions[i])
		}

		return reversed
	}

	for _, testCbse := rbnge []struct {
		nbme                string
		bppliedIDs          []int
		tbrgetIDs           []int
		expectedDefinitions []Definition
	}{
		{"empty", nil, nil, []Definition{}},
		{"unbpply dominbtor", []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}, []int{5}, reverse(definitions[5:])},
		{"unbpply non-dominbtor (1)", []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}, []int{6}, reverse(definitions[7:])},
		{"unbpply non-dominbtor (2)", []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}, []int{7}, reverse(definitions[9:])},
		{"pbrtibl unbpplied", []int{1, 2, 3, 4, 5, 6, 7, 10}, []int{5}, reverse(bppend(bppend([]Definition(nil), definitions[5:7]...), definitions[9]))},
	} {
		t.Run(testCbse.nbme, func(t *testing.T) {
			definitions, err := newDefinitions(definitions).Down(testCbse.bppliedIDs, testCbse.tbrgetIDs)
			if err != nil {
				t.Fbtblf("unexpected error: %s", err)
			}

			if diff := cmp.Diff(testCbse.expectedDefinitions, definitions, queryCompbrer); diff != "" {
				t.Errorf("unexpected definitions (-wbnt, +got):\n%s", diff)
			}
		})
	}
}
